Revert "compatibility"
This reverts commit 45da2936368dc899a2c3eecc75381d2e08d96a8c.

+++ segpipe/pipeline.py
@@ -73,7 +73,7 @@ class segmentationPipeline:
         LRInput = torchPrep(lrImage) #HWD -> NCHWD
         LROutput = torchGetModelOutput(LRInput,self.LRModel)
         LROutput = torchDust(LROutput,threshold=5000,takeLargest=takeLargest)
-        LROutput = torch.nn.functional.interpolate(LROutput.float(), size=originalImage.shape[2:], mode='nearest').int()
+        LROutput = torch.nn.functional.interpolate(LROutput, size=originalImage.shape[2:], mode='nearest')
         
 
         if getLR == 1:
@@ -119,7 +119,7 @@ class segmentationPipeline:
             # Get and post-process left lobe model output
             leftInput = torchPrep(leftCropped)
             leftLobeOutput = torchGetModelOutput(leftInput,self.leftModel)
-            leftLobeOutput = torch.nn.functional.interpolate(leftLobeOutput.float(), size=leftCropped.shape[2:], mode='nearest').int()
+            leftLobeOutput = torch.nn.functional.interpolate(leftLobeOutput, size=leftCropped.shape[2:], mode='nearest')
             temp = torch.zeros_like(leftLobeOutput)
             temp[:,:,:-1,:-1,:-1] = leftLobeOutput[:,:,1:,1:,1:]
             leftLobeOutput = temp
@@ -127,7 +127,7 @@ class segmentationPipeline:
             # Get and post-process right lobe model output
             rightInput = torchPrep(rightCropped)
             rightLobeOutput = torchGetModelOutput(rightInput,self.rightModel)       
-            rightLobeOutput = torch.nn.functional.interpolate(rightLobeOutput.float(), size=rightCropped.shape[2:], mode='nearest').int()
+            rightLobeOutput = torch.nn.functional.interpolate(rightLobeOutput, size=rightCropped.shape[2:], mode='nearest')
 
             #adjust right lobe output (0,1,2,3) to (0,3,4,5)
             rightLobeOutput = rightLobeOutput + 2
@@ -159,13 +159,13 @@ class segmentationPipeline:
         for i in range(3):
             if unevenShape[i]:
                 shape[i+2] += 1
-        finalMask = torch.nn.functional.interpolate(finalMask.float(), size=shape[2:], mode='nearest-exact').int()
+        finalMask = torch.nn.functional.interpolate(finalMask, size=shape[2:], mode='nearest-exact')
         if postprocess:
             finalMask = torchErrors(finalMask)
         finalMask = torchDust(finalMask)
         #finalMask = torchSmoothing(finalMask)
 
-        finalMask = torch.nn.functional.interpolate(finalMask.float(), size=originalImage.shape[2:], mode='nearest-exact').int()
+        finalMask = torch.nn.functional.interpolate(finalMask, size=originalImage.shape[2:], mode='nearest-exact')
         
         finalMask = finalMask.to(torch.uint8)
